requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.26
scipy>=1.11
opencv-python>=4.9
pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
pandas>=2.1
matplotlib>=3.8
scikit-learn>=1.4
tensorboard>=2.15
einops>=0.7
timm>=1.0
